requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
pytz==2022.5
sqlparse==0.4.3
xlrd==2.0.1
